refactor: rename events API to activities with recordings endpoint
- Rename events router package to activities
- Update endpoint from /api/events to /api/activities/recordings
- Change function name from get_events to get_recording
- Update frontend to use new API endpoint
- Remove old events directory and files
