+++ getgather/api/main.py
@@ -14,7 +14,7 @@ from jinja2 import Template
 
 from getgather.api.routes.auth.endpoints import router as auth_router
 from getgather.api.routes.brands.endpoints import router as brands_router
-from getgather.api.routes.events.endpoints import router as events_router
+from getgather.api.routes.activities.endpoints import router as activities_router
 from getgather.api.routes.link.endpoints import router as link_router
 from getgather.browser.profile import BrowserProfile
 from getgather.browser.session import BrowserSession
@@ -235,6 +235,6 @@ async def extended_health():
 
 app.include_router(brands_router)
 app.include_router(auth_router)
-app.include_router(events_router)
+app.include_router(activities_router)
 app.include_router(link_router)
 app.mount("/mcp", mcp_app)
